Add PAD_CFG_GPI_INT(pad, pull, rst, trig) macro
This macro should be used in cases when it is necessary to set the
trig parameter. The following entry is often found in inteltool logs:

0x0428: 0x0000004d84000100 GPP_C5   GPIO

RX Level/Edge Configuration set as Drive '0' (2h) [1]. This is also
the default value for this field in the DW0 register. Since standard
macros are preferable for use, only PAD_CFG_GPI_INT() can set the
value of this field by default:

PAD_CFG_GPI_INT(GPP_C5, NONE, PLTRST, OFF), /* GPIO */

[1] p1431, Intel 100/C230 Series Chipset Family PCH Volume 2 (February
2019)

Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>