Set -raw command line option instead of unused -gdb
The -raw flag forces the generation of common macros with raw registers
DW0, DW1 values:

  _PAD_CFG_STRUCT(GPP_A12, 0x84000102, 0x00000024), /* GPIO */
  _PAD_CFG_STRUCT(GPP_A13, 0x44000600, 0x00000025), /* SUSWARN# */

Thus no longer need to create separate files to generate/ directory.
In addition to this, the patch removes the -dbg option, since there
are no plans to use it. All logs are printed to the console.

Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>