parser: use new tamplates for parsing config file
This patch makes changes to the decoding process of the configuration
file lines. Now the parser uses templates to decode information. The
template type matches the format of the configuration file and it is
set using command line options -t <type num>. At the moment, it is
possible to parse inteltool.log and gpio.h (file from the coreboot/
src/mainboard/<board> configuration with sunrise PCH). Also, the user
can set his own template in parser/template and run the utility with
the key "-t 2".

Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>