common: add a package that contains common code for all platforms
The utility supports several platforms that have a common code part.
Thus, in order to avoid code duplication, this patch adds the "common"
package, which allows to union repeating functions from all other
platform packages. To take into account platform-specific functions,
a special interface is defined inside the new package:

type PlatformSpecific interface {}

The functions in this interface should be defined in a package that
uses a common part of the code:

  - Adds the PADRSTCFG parameter from PAD_CFG_DW0 to the macro as a
    new argument:
        func Rstsrc(macro *Macro) {}

  - Adds The Pad Termination (TERM) parameter from PAD_CFG_DW1 to the
    macro as a new argument:
        func Pull(macro *Macro) {}

  - Adds PAD_CFG_GPI macro with arguments:
        func GpiMacroAdd(macro *Macro) {}

  - Adds PAD_CFG_GPO macro with arguments:
        func GpoMacroAdd(macro *Macro) {}

  - Adds PAD_CFG_NF macro with arguments:
        func NativeFunctionMacroAdd(macro *Macro) {}

  - Generates "Advanced" macro if the standard ones can't define the
    values from the DW0/1 register:
        func AdvancedMacroGenerate(macro *Macro) {}

Thus, the layer-by-layer separation architecture is implemented.

Now the term "Advanced" is used instead of "Common" for macros to avoid
confusion with the name of the new package. Also it should be renamed in
all other packages, regardless of the platform type and the presence
of common code in the file.

Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>