Revert "common/macro: undo checking mask for NF macro"
This reverts commit 8ac6861c53d1a4d9357731b5531bd040a34a148d.

The utility should check all fields of the DW0 register. If a register
field is found that was not used to generate the macro, the utility will
generate an advanced macro, and the user must take care of this register
field on his own. If the user decides to ignore this field, then he must
clear this value in the dump of the inteltool registers, and then run
the parser utility again.

Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>